[Feature #153114603]Move repeated method to mixin class
