+++ hc/blog/mixins.py
@@ -1,3 +1,12 @@
+"""
+Module with custom mixin classes to avoid code repetition.
+"""
+
+from django.contrib import messages
+from .models import Post
+from braces.views import UserPassesTestMixin
+
+
 class CommonContentMixin(object):
     """
     Defines attributes that are common or needed by almost all views.
@@ -9,3 +18,32 @@ class CommonContentMixin(object):
         if previous_url:
             ctx['previous_url'] = previous_url
         return ctx
+
+
+class PostOwnerRequiredMixin(UserPassesTestMixin):
+    """
+    Handle views that require users to be the owners of the blog post objects.
+
+    Users who require to perform delete and update actions must pass the test method below.
+    """
+
+    def test_func(self, user):
+        """
+        Define rules that user must pass
+
+        :param user: logged in user
+        :return: bool
+        """
+        slug = self.kwargs.get('slug', None)
+        if slug:
+            obj = Post.objects.filter(slug=slug)
+
+            if obj.exists():
+                owner = obj.first().created_by == user
+
+                if not owner:
+                    messages.warning(self.request, 'You are not the owner of the blog post!')
+
+                return owner
+
+        return False
